refactor: command interface to cli pkg
