+++ main.go
@@ -5,6 +5,7 @@ import (
 	"os"
 
 	_ "github.com/mattn/go-sqlite3"
+	"github.com/nurali-techie/rc/cli"
 	"github.com/nurali-techie/rc/command"
 	"github.com/nurali-techie/rc/config"
 	"github.com/nurali-techie/rc/database"
@@ -34,7 +35,7 @@ func main() {
 	// init commander
 	helpCmd := command.NewHelpCommand(console)
 	getCmd := command.NewGetCommand(commandService, clipboard)
-	commander := NewCommander(helpCmd, getCmd)
+	commander := cli.NewCommander(helpCmd, getCmd)
 
 	// register commands
 	commander.Register("add", command.NewAddCommand(commandService, clipboard))
